refactored abstract base class
