+++ autoscaler/modes/scalemem.py
@@ -1,10 +1,8 @@
 
-from autoscaler.modes.scalemode import AbstractMode
+from autoscaler.modes.abstractmode import AbstractMode
 
 
 class ScaleByMemory(AbstractMode):
-
-    MODE_NAME = 'mem'
 
     def __init__(self, api_client=None, app=None, dimension=None):
         super().__init__(api_client, app, dimension)
@@ -43,18 +41,7 @@ class ScaleByMemory(AbstractMode):
         if value == -1.0:
             return 0
 
-        if value > self.get_max():
-            self.log.info("%s above max threshold of %s"
-                          % (self.MODE_NAME, self.get_max()))
-            return 1
-        elif value < self.get_min():
-            self.log.info("%s below threshold of %s"
-                          % (self.MODE_NAME, self.get_min()))
-            return -1
-        else:
-            self.log.info("%s within thresholds (min=%s, max=%s)"
-                          % (self.MODE_NAME, self.get_min(), self.get_max()))
-            return 0
+        return super().scale_direction(value)
 
     def get_mem_usage(self, task, agent):
         """Calculate memory usage for the task on the given agent
